requirements.txt
# Runtime dependencies for all guardrail modules
flask>=3.0.0
gunicorn>=21.2.0
prometheus-client>=0.20.0
PyYAML>=6.0

# Optional LLM backends (install as needed)
# openai>=1.0.0
# anthropic>=0.20.0

# Dev / test
pytest>=8.0.0
pytest-cov>=5.0.0
